updated the readme to give better install instructions
I needed to update this readme after having some issues with a grader giving me a failing grade because they, themselves didn't know how to connect to their MySQL server, even though my repo in its current state not only works on my machine but also another machine with their own SQL user/password too. I understand using dotenv can be confusing and maybe the way I wrote it was confusing I guess since I didn't include an env template? Even though the way I wrote it was providing a js file that double-checked the variables to see if they were corrected and then also implemented a triple check to see if you typed in the variables correctly on the main js file. So if you didn't have a template but looked in the connection.js file (which was the file I am referencing) and assuming you knew how dotenv worked you can at least get a good, crystal clear idea of what the .env variables were called and just write them out for connecting to your own server. Which yeah I get is a lot of work but when I also give the said person the exact .env file that I personally used and change my whole repo to bypass the dotenv stuff entirely and they somehow get the same issue regarding the dotenv stuff... yeah that's when I realized I should probably give this an update. Got a 100 on this though which is cool but yeah if you didn't know your MySQL server username/password would you think it was on you or the person who made it? lol thanks for reading this commit